import writer init

+++ src/main.ts
@@ -2,8 +2,10 @@ import { Product } from './product';
 import { LeroyScraper } from "./scraper/leroy-scraper";
 import * as fs from 'fs';
 import { auditTime } from 'rxjs';
+import { writeImports } from './import-writer';
+import { ScraperResult } from './scraper/scraper';
 
-async function main() {
+const runScraper = async (): Promise<ScraperResult> => {
   const scraper = new LeroyScraper();
 
   const sub = scraper.progressObs
@@ -24,9 +26,30 @@ async function main() {
     });
   
   const result = await scraper.scrapeWebsite();
+  return result;
+} 
 
-  console.log("Saving results.");
-  fs.writeFileSync('./out/out.json', JSON.stringify(result));
+async function main() {
+  if(!fs.existsSync('./out/images')) {
+    fs.mkdirSync('./out/images', {recursive: true});
+  }
+  if(!fs.existsSync('./out/imports')) {
+    fs.mkdirSync('./out/imports', {recursive: true});
+  }
+
+  let result: ScraperResult;
+  if(process.argv.length > 2 && process.argv[2] === 'imports' && fs.existsSync('./out/out.json')){
+    result = JSON.parse(fs.readFileSync('./out/out.json').toString());
+  }
+  else {
+    result = await runScraper();
+
+    console.log("Saving results.");
+    fs.writeFileSync('./out/out.json', JSON.stringify(result));
+  }
+
+  console.log('Writing imports.');
+  writeImports(result);
 };
 
 main();
